Extract secp256k1 adapter tests.
